Napravljen dodavanje racuna sa stavkama u bazu podataka.

+++ Software/ZMGDesktop/EntitiesLayer/Entities/StavkaRacun.cs
@@ -33,7 +33,7 @@ namespace EntitiesLayer.Entities
 
         public int KolikoRobePoJedinici { get; set; }
 
-        public float JedinicnaCijena { get; set; }
+        public double JedinicnaCijena { get; set; }
 
         public DateTime? DatumIzrade { get; set; }
